requirements.txt
langchain
langchain-google-genai
python-dotenv
pdfplumber
langchain-google-vertexai
faiss-cpu
google-auth
